Define fallbackMusic level property to make music dlc possible Move autotest ai to separate file
