+++ RemixedDungeon/build.gradle
@@ -28,8 +28,8 @@ android {
                                 "gow_app_id":getApiField("gow_app_id")]
 
         applicationId "com.nyrds.pixeldungeon.ml"
-        versionCode 957
-        versionName "29.6.beta.1"
+        versionCode 958
+        versionName "29.6.beta.2"
 
         buildToolsVersion = '29.0.2'
         targetSdkVersion 29
