+++ RemixedDungeon/src/main/assets/levelsDesc/Dungeon_debug.json
@@ -78,9 +78,9 @@
     "32":{"kind":"LastLevel", "depth":32,       "size":[32,32]}
     },
     "Graph":{
-    "0":[["1"],[]],
+    "0":[["town_2"],[]],
     "testLevel":[[],[]],
-    "random":[[],[]],
+    "random":[["1"],["1"]],
 
     "town_2":[["church_2","cinema_2","1","library_2","shop_town_2","fortune_2"],[]],
     "church_2":[[],["town_2"]],
