+++ RemixedDungeon/src/main/assets/scripts/services/autoTestAi.lua
@@ -1,0 +1,105 @@
+---
+--- Generated by EmmyLua(https://github.com/EmmyLua)
+--- Created by mike.
+--- DateTime: 12/23/20 10:01 PM
+---
+
+local RPD = require "scripts.lib.commonClasses"
+
+local ai = {}
+
+ai.step = function()
+    local hero = RPD.Dungeon.hero
+
+    if not hero:isReady() then
+        return
+    end
+
+    local activeWindow = RPD.RemixedDungeon:scene():getWindow(0)
+
+    if activeWindow then
+
+        local wndClass = tostring(activeWindow:getClass())
+
+        if wndClass:match('WndInfoMob') then
+            local target = activeWindow:getTarget()
+            local action = RPD.CharUtils:randomAction(target,hero)
+            RPD.CharUtils:execute(target, hero, action);
+        end
+
+        if wndClass:match('WndStepOnTrap') then
+            activeWindow:onSelect(0)
+        end
+
+        if wndClass:match('WndChasmJump') then
+            if math.random()<0.05 then
+                activeWindow:onSelect(0)
+            else
+                activeWindow:onSelect(1)
+            end
+        end
+
+        activeWindow:hide()
+        return
+    end
+
+    local level = hero:level()
+
+    if hero:buffLevel('Blindness') > 0 then
+        local cell = level:getEmptyCellNextTo(hero:getPos())
+        if level:cellValid(cell) then
+            hero:handle(cell)
+        else
+            hero:rest(false)
+        end
+        return
+    end
+
+    local enemyPos = hero:getNearestEnemy():getPos()
+    if level:cellValid(enemyPos) then
+        hero:handle(enemyPos)
+        return
+    end
+
+    if hero:buffLevel('Roots') > 0 or math.random() < 0.1 then
+        hero:search(true)
+        return
+    end
+
+    local exitCell = level:getRandomVisibleTerrainCell(RPD.Terrain.EXIT)
+
+    if level:cellValid(exitCell) and not level:getTopLevelObject(exitCell) then
+        hero:handle(exitCell)
+        return
+    end
+
+    exitCell = level:getRandomVisibleTerrainCell(RPD.Terrain.LOCKED_EXIT)
+    if level:cellValid(exitCell) and not level:getTopLevelObject(exitCell) and hero:getBelongings():getItem("SkeletonKey"):valid() then
+        hero:handle(exitCell)
+        return
+    end
+
+    local doorCell = level:getRandomVisibleTerrainCell(RPD.Terrain.DOOR)
+
+    if level:cellValid(doorCell) and  not level:isCellVisited(doorCell) then
+        hero:handle(doorCell)
+        return
+    end
+
+    doorCell = level:getRandomVisibleTerrainCell(RPD.Terrain.LOCKED_DOOR)
+
+    if level:cellValid(doorCell) and hero:getBelongings():getItem("IronKey"):valid() then
+        hero:handle(doorCell)
+        return
+    end
+
+
+    local cell = level:randomTestDestination()
+    if not level:cellValid(cell) then
+        cell = level:randomDestination()
+    end
+
+    hero:handle(cell)
+end
+
+return ai
