+++ build.gradle
@@ -7,8 +7,8 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:4.1.1'
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.3.0'
-        classpath 'com.google.firebase:perf-plugin:1.3.3'
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.4.1'
+        classpath 'com.google.firebase:perf-plugin:1.3.4'
         classpath 'com.google.gms:google-services:4.3.4' // google-services plugin
         classpath 'com.getkeepsafe.dexcount:dexcount-gradle-plugin:1.0.0'
         classpath 'com.huawei.agconnect:agcp:1.4.1.300'
